Adding new folders names, and creating new UsualCustomers.jsx components too
