+++ frontend/src/pages/parkingSpace/ParkingStatistics.jsx
